fix: rename components folder
