+++ src/app/(main)/student/browse/_components/SearchBar.tsx
